[bugfix] update to clean drivers api for mpsize

+++ api/socs/stm32f439/usbctrl_backend.h
@@ -189,7 +189,7 @@ mbed_error_t usb_backend_drv_stall(uint8_t ep_id, usb_backend_drv_ep_dir_t dir);
 mbed_error_t usb_backend_drv_endpoint_disable(uint8_t ep_id, usb_backend_drv_ep_dir_t dir);
 mbed_error_t usb_backend_drv_endpoint_enable(uint8_t ep_id, usb_backend_drv_ep_dir_t dir);
 
-uint32_t usb_backend_get_ep_mpsize(void);
+uint16_t usb_backend_get_ep_mpsize(void);
 
 usb_backend_drv_port_speed_t usb_backend_drv_get_speed(void);
